fix database-config and add-migration to sql-server

+++ WAAL.Domain/Entities/PhieuNhap.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public DateTime NgayNhap { get; set; }
-        public double TongTien { get; set; }
+        public decimal TongTien { get; set; }
         public string TinhTrang { get; set; }
         public bool TrangThai { get; set; }
         public NhanVien NhanVien { get; set; }
